backend/requirements.txt
Flask
Flask-CORS
Werkzeug
Pillow
torch
torchvision
transformers
accelerate
pydantic
requests
hf_transfer
python-dotenv

